refactor: extract selection controller from LdSelectableList
- Extract LdSelectableListSelectionController into separate file
- Reduce selectable_list.dart from 777 to 508 lines
- Controller manages all selection state with ChangeNotifier pattern
- Preserve Stage 1 dispose() logic in controller's dispose method
